enhancement: Generate OptimizelyConfig object on API Call instead of SDK initialization (#296)
## Summary
`OptimizelyConfig` object is generated when SDK is being initialized. This operation is heavy and blocks the SDK from initializing quickly. This object is only used when users explicitly access it using `get_optimizely_config` API call. This PR moves the creation of this object from SDK initialization to the actual API call.

## Test plan
- Manually tested thoroughly
- All unit tests pass
- All FSC tests pass

## Issues
fixes #295